fix(ui): use `sessionStorage` instead of `localStorage`
It prevents some settings to be shared accross multiples apps on the
same domain
